restrukturisasi code test3
